+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [unreleased]
 
+### Added
+
+- Add send token to addresses in `HFunds`.
+- Support swap of Stargate token bridge.
+- Add `setTokenBalance` in utils.js to manipulate token balance for unit testing.
+
 ## [2.3.2] - 2022-12-21
 
 ### Changed